fixed and improved
Fixed: When for a company with no templates, if a template is first
added to success table and then when another template is not being added
to failure db.

Improved: No multiple commits when a new template file is added for a
company which does not exist. Previously there were two commits is being
done in this case. First commit when the company is added to the db and
second commit is adding the template under that company.